feat(internal-economics): promote presentation rounding oracle to shared lib (D1)
Promote the hierarchical exact-Decimal LRM rounding oracle from the
truth-case helper into shared/lib/internal-economics/presentation-rounding-v1
(byte-identical modulo the Task163 name prefix). The helper becomes a thin
re-export shim preserving the Task163-prefixed aliases so the pinned
readiness test is unchanged.
